add delay effect :sparkles:

+++ pkg/dsp/window/bartlett.go
@@ -2,11 +2,11 @@ package window
 
 import "fmt"
 
-func Bartlett(len int) []float32 {
+func Bartlett(len int) []int32 {
 
-	output := make([]float32, len)
+	output := make([]int32, len)
 
-	N := float32(len - 1)
+	N := len - 1
 
 	coefficients := 2 / N
 	n := 0
@@ -14,11 +14,11 @@ func Bartlett(len int) []float32 {
 	fmt.Println(N)
 
 	for ; n <= N/2; n++ {
-		output[n] = float32(n * int(coefficients))
+		output[n] = int32(n * coefficients)
 	}
 
 	for ; n <= N; N++ {
-		output[n] = 2 - float32(n*int(coefficients))
+		output[n] = int32(2 - n*int(coefficients))
 	}
 
 	return output
